:truck: seperate out flat list item component
