+++ src/screens/currency-list/flat-list-item/index.js
@@ -1,0 +1,1 @@
+export { FlatListItem } from './flat-list-item';
