+++ src/screens/currency-list/flat-list-item/prop-types.js
@@ -1,0 +1,9 @@
+import { bool, func, shape, string } from 'prop-types';
+
+export const propTypes = {
+  item: string.isRequired,
+  isBaseCurrency: bool.isRequired,
+  navigation: shape({
+    pop: func.isRequired,
+  }).isRequired,
+};
